requirements.txt
fastapi>=0.100
uvicorn>=0.22
sentence-transformers>=2.2
networkx>=3.1
numpy>=1.24
httpx>=0.24
scipy>=1.11
pytest>=7.4
pytest-asyncio>=0.21
pytest-cov>=4.1
python-dotenv>=1.0
